pyproject.toml
[project]
name = "crewai-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "crewai-tools>=0.45.0",
    "fastapi>=0.115.12",
    "fastmcp>=2.5.2",
    "langchain-groq>=0.3.2",
    "mcp-use>=1.3.0",
    "mcp[cli]>=1.9.2",
]

[project.scripts]
blog_creator = "blog_creator.main:run"
run_crew = "blog_creator.main:run"
train = "blog_creator.main:train"
replay = "blog_creator.main:replay"
test = "blog_creator.main:test"


[tool.crewai]
type = "crew"
